Slingshot Attack Montage
Since the 'Assets' directory is ignored, the following files are not committed/pushed:

[1] Content/Assets/Enemies/Goblin/Animations/Slingshot/AM_Attack_Goblin_Slingshot.uasset
Added a new Notify Track, called "Rock", and two new Skeleton Notifies on that track, called 'ReleaseRock' and 'GrabPouch'.

[2] Content/Assets/Enemies/Goblin/Slingshot/Centered/AM_Slingshot_Attack.uasset
Created new Animation Montage for slingshot attack animation.

[3] Content/Assets/Enemies/Goblin/SKM_Goblin.uasset
[4] Content/Assets/Enemies/Goblin/SK_Goblin.uasset
Changed the Preview Asset to "SKM_Slingshot". This does not impact with game logic but still it required to save the file.

Extra Notes:

[1] Content/Blueprints/Character/Aura/BP_AuraCharacter.uasset
This file is not committed because nothing is actually changed here.
The original course repository have a commit for this file because it changed the "Target Arm Length" for `SpringArm` component and later it restored back to its previous value.