Decrementing Minion Count
Extra Notes:

[1] Content/Blueprints/AbilitySystem/Enemy/Abilities/GA_SummonAbility.uasset
Changed "Instancing Policy" of this gameplay ability to 'Instanced Per Actor'. Otherwise, it does not spawn the minions when the ability is ended. We can avoid calling the 'End Ability' function to make it work, but it may not be the best practice to keep the ability instance in memory once it's done executing.

So, instead of using the "Instancing Policy" of 'Instanced Per Execution' without calling the 'End Ability' function, it is better to use 'Instanced Per Actor' with the call to the 'End Ability' function to make sure of optimal usage.

When using the "Instancing Policy" of 'Instanced Per Actor', we manually need to take care of resetting the variables used within that ability to their default values, or they will use the values from previous executions for the actor who is using that ability.