Adding Juice with Tweening
Extra Notes:

[1] Content/Blueprints/Character/BP_EnemyBase.uasset
Set the 'Length' of `SpawnTimeline` node to value 0.3 instead of default 5.0. This is to trim the unneeded time after the last key on the track, so it ticks just what it is necessary.
This may not be a big deal in the current context but it can always add up and it's a good practice to keep timelines tight.