Fade Actor
Since the 'Assets' directory is ignored, the following files are not committed/pushed:

[1] Content/Assets/Dungeon/Beacon/Material/M_Beacon_f.uasset
This is a duplicate material made from `M_Beacon`, and it sets the 'Blend Mode' to "Masked". It also makes use of 'Opacity Mask' parameter of material's result node by providing a value coming from DitherTemporalAA node that uses 'Fade' parameter.

[2] Content/Assets/Dungeon/Beacon/Material/M_Beacon_f_Inst.uasset
Instance of above material that can use 'Fade' and 'Glow' parameters. This is used by `BP_FadeActor` to set its 'FadeMaterialInstances' variable.