Spell Menu Button
Since the 'Assets' directory is ignored, the following files are not committed/pushed:

[1] Content/Assets/UI/Globes/MI_FlowingUIBG.uasset
This material instance is updated with values for Global Scalar Parameter Values and Global Vector Parameter Values.

[2] Content/Assets/UI/Globes/MI_Flowing_Scroll.uasset
This material instance is updated with values for Global Scalar Parameter Values and Global Vector Parameter Values.

[3] Content/Assets/UI/Globes/M_FlowingUIBG.uasset
This material is updated with 'UTiling' and 'VTiling' parameter values for both "TexCoord[0]" nodes.