Debuff Niagara Component
Extra Edits:

[1] Source/Aura/Public/Interaction/CombatInterface.h
Used the reference of `FOnDeath` delegate as a return type of the pure virtual function `GetOnDeathDelegate()`. This is to avoid returning copy of that delegate which is not useful to respond with respect to actual delegate.

[2] Source/Aura/Public/Character/AuraCharacterBase.h
[3] Source/Aura/Private/Character/AuraCharacterBase.cpp
Used the reference of `FOnDeath` delegate as a return type while overriding the pure virtual function `GetOnDeathDelegate()`.